+++ src/services/sesionesService.ts
@@ -57,14 +57,11 @@ export const obtenerInformeAsistencia = async (
     .then((response) => response.data);
 };
 
-export const obtenerAllSesionesPorMes = async (
-  mes: number,
-  anio: number
-) => {
+export const obtenerAllSesionesPorMes = async (mes: number, anio: number) => {
   return api
     .get(`talleres-sesiones/asistencia/all/${mes}/${anio}`)
     .then((response) => response.data);
-}
+};
 
 export const obtenerAllSesionesPorMesDetalle = async (
   mes: number,
@@ -72,5 +69,22 @@ export const obtenerAllSesionesPorMesDetalle = async (
 ) => {
   return api
     .get(`talleres-sesiones/asistencia/all-detalles/${mes}/${anio}`)
+    .then((response) => response.data);
+};
+
+export const obtenerPorcentajeAsistencia = async (
+  mes: number,
+  anio: number
+) => {
+  return api
+    .get(`talleres-sesiones/porcentaje-asistencia/${mes}/${anio}`)
+    .then((response) => response.data);
+};
+
+export const obtenerEstadisticasAsistencia = async (
+
+) => {
+  return api
+    .get(`talleres-sesiones/porcentaje-asistencia/general`)
     .then((response) => response.data);
 }